% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.R
\name{iterate}
\alias{iterate}
\alias{iterate,lhm,numeric-method}
\title{Add iterations to existing \code{lhm} object}
\usage{
iterate(object, iter, ...)

\S4method{iterate}{lhm,numeric}(object, iter)
}
\arguments{
\item{object}{existing \code{lhm} object}

\item{iter}{number of iterations}

\item{...}{arguments to generic function}
}
\value{
Returns an \code{lhm} object with life-history data replicated \code{iter} times
}
\description{
This function will add an iteration dimension to the life-history data contained in an \code{lhm} object.
}
\examples{
# initialise data object
dat <- lhm(ainf = 30, iter = 1)

# to generate monte-carlo iterations first either 
# re-initialise 'lhm' object using more iterations or 
# iterate existing object:
dat <- iterate(dat, iter = 200)

}
