% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration.R
\name{iteration}
\alias{iteration}
\alias{iteration,lhm,numeric-method}
\title{Extract single iteration from \code{lhm} object}
\usage{
iteration(object, iter, ...)

\S4method{iteration}{lhm,numeric}(object, iter)
}
\arguments{
\item{object}{an \code{lhm} object}

\item{iter}{iteration to be extracted}

\item{...}{arguments to generic function}
}
\value{
An object of class \code{\link{lhmIter-class}} which contains a single iteration only.
}
\description{
Extracts a single iteration from an \code{lhm} object.
}
