% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhm-class.R, R/lhm.R
\docType{class}
\name{lhm-class}
\alias{lhm-class}
\alias{lhm}
\title{Class containing life-history information}
\usage{
lhm(ainf, iter = 1, ...)
}
\arguments{
\item{ainf}{the assumed asymptotic age}

\item{iter}{the number of iterations}

\item{...}{additional arguments \code{nmort}, \code{size}, \code{mass}, \code{sr} and \code{maturity} that can be optionally specified at initialisation of the object}
}
\description{
This is an S4 object class that contains the life-history information necessary to calculate derived quantities using life history theory, such as the intrinsic growth rate \eqn{r} using the Euler-Lotka equation. 

Initialisation is usually through a call to \code{lhm(ainf,iter)}. Using this approach specific life-history data can then be included through the assignment functions \code{\link{nmort}}, \code{\link{size}}, \code{\link{mass}}, \code{\link{sr}} and \code{\link{maturity}}.
}
\section{Slots}{

\describe{
\item{\code{iter}}{the number of Monte-Carlo iterations}

\item{\code{ainf}}{the asymptotic age (equal to the maximum age for each life-history data vector)}

\item{\code{sr}}{the type of stock-recruitment relationship used, either Beverton-Holt (\code{'BH'}) or Ricker (\code{'RK'})}

\item{\code{lhdat}}{list of data matrices containing life-history data-at-age, specifically, mass, maturity, natural mortality and survivorship, as well as a vector of steepness values}
}}

\examples{
# initialise lhm-class object
ainf <- 100
iter <- 1
dat <- lhm(ainf, iter)

# assign life-history data
nmort(dat)    <- list(mu=0.18)
size(dat)     <- list(mu=list(Linf=106.5,k=0.229,t0=0.01)) 
mass(dat)     <- list(mu=list(a=1.7e-9,b=3.328))
sr(dat)       <- list(type='BH',mu=0.9)
maturity(dat) <- list(mu=list(acrit=8))

}
\seealso{
See the appropriate documentation in \code{\link{mass}}, \code{\link{maturity}}, \code{\link{nmort}}, \code{\link{size}}, \code{\link{sr}} and \code{\link{survivorship}} for how these are specified, accessed and assigned. See documentation for \code{\link{rCalc}} on how to calculate \eqn{r} from an \code{lhm} object. Use \code{\link{iterate}} to increase the number of iterations after the object has been created. Use \code{\link{iteration}} to access a particular iteration.
}
