% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhmIter-class.R, R/lhmIter.R
\docType{class}
\name{lhmIter-class}
\alias{lhmIter-class}
\alias{initialize,lhmIter-method}
\title{Class containing single iteration of life-history information}
\usage{
\S4method{initialize}{lhmIter}(.Object, ainf, sr)
}
\arguments{
\item{.Object}{object for retun}

\item{ainf}{the assumed asymptotic age}

\item{sr}{stock recruitment function}
}
\description{
This is an S4 object class similar to \code{\link{lhm}}. It differs from \code{lhm} in that life-history data are stored as vectors (rather than matrices) and therefore correspond to a single monte-carlo interation only. The vectorisation is necessary for efficient estimation of derived quantities such as \eqn{r} using \code{\link{rCalc}}.

Initialisation is through a call to \code{\link{iteration}}, which will extract a single iteration from an \code{lhm} object.
}
\section{Slots}{

\describe{
\item{\code{ainf}}{the asymptotic age (equal to the maximum age for each life-history data vector)}

\item{\code{sr}}{the type of stock-recruitment relationship used, either Beverton-Holt (\code{'BH'}) or Ricker (\code{'RK'})}

\item{\code{lhdat}}{list of numeric vectors containing life-history data-at-age, specifically, mass, maturity, natural mortality and survivorship, as well as a steepness value}
}}

\seealso{
\code{\link{lhm}}, \code{\link{iteration}}, \code{\link{rCalc}}
}
