% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmort.R
\name{nmort}
\alias{nmort}
\alias{nmort,lhm-method}
\alias{nmort<-}
\alias{nmort<-,lhm,list-method}
\alias{nmort<-,lhm,numeric-method}
\title{Access or assign natural mortality at age}
\usage{
nmort(object, ...)

\S4method{nmort}{lhm}(object)

nmort(object) <- value

\S4method{nmort}{lhm,list}(object) <- value

\S4method{nmort}{lhm,numeric}(object) <- value
}
\arguments{
\item{object}{a \code{\link{lhm}} object}

\item{...}{arguments to generic function}

\item{value}{either a named \code{list} or \code{numeric} vector 
\describe{
 \item{\code{numeric}}{a single numeric or vector of values}
 \item{\code{list}}{a list object containing the components \code{mu}, \code{cv} and \code{range}}
 \item{\code{mu}}{a single numeric or vector of values representing the mean natural mortality}
 \item{\code{cv}}{optional single numeric value representing the coefficient of variation}
 \item{\code{range}}{optional vector of length two containing the lower and upper bounds}
}}
}
\value{
Accessor function returns a matrix of maturity at age across iterations. Assignment function populates the matrix across iterations.
}
\description{
This function can be used to access or assign the natural mortality at age in an \code{\link{lhm}} object. 

Assignment can be of a list or numeric vector. When a list is assigned it should contain a value for \code{mu}, which is the mean natural mortality at age and may or may not be constant across ages.

If the list also contains \code{cv} then log-normal uncertainty is added around the mortality at age. If the list contains \code{range} then the uncertainty is bounded at these values. If \code{range} is provided but not \code{cv}, and if mortality is constant across ages, then a uniform uncertainy is implemented.
}
\examples{
# M at age vector input
dat <- lhm(ainf = 10,iter=1)
nmort(dat) <- c(0.1,0.1,0.2)
nmort(dat)

# constant M with log-normal uncertainty
dat <- lhm(ainf=10, iter = 20)
nmort(dat) <- list(mu = 0.2, cv = 0.25)
nmort(dat)

# constant M with bounded log-normal uncertainty
dat <- lhm(ainf = 10, iter = 20)
nmort(dat) <- list(mu = 0.2, cv = 0.5, range = c(0.1,0.3))
nmort(dat)

# constant M with uniform uncertainty
dat <- lhm(ainf = 10,iter = 20)
nmort(dat) <- list(range = c(0.1,0.3))
nmort(dat)

# M at age with log-normal uncertainty
dat <- lhm(ainf=10, iter = 20)
nmort(dat) <- list(mu = c(0.1,0.1,0.2), cv = 0.25)
nmort(dat)

# M at age with bounded log-normal uncertainty
dat <- lhm(ainf = 10, iter = 20)
nmort(dat) <- list(mu = c(0.1,0.1,0.2), cv = 0.25, range = c(0.1,0.3))
nmort(dat)

}
