% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.prior.R
\name{plot.prior}
\alias{plot.prior}
\title{Plot prior distribution}
\usage{
\method{plot}{prior}(x, ...)
}
\arguments{
\item{x}{an \code{\link{prior}} object class}

\item{...}{additional arguments for \code{\link[graphics]{hist}}}
}
\description{
This function provides a simple \code{plot} method for the \code{prior} object class.
}
\examples{
# create object containing
# vector of r values
iter <- 1000
mu <- 0.1
cv <- 0.2
sd <- sqrt(log(1+cv^2))
x <- rlnorm(iter,log(mu)-sd^2/2,sd)
r <- prior(x)

# plot
\dontrun{
plot(r, xlab = 'r', ylab = '', yaxt = 'n', main = '')
}

}
