% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCalc.R
\name{rCalc}
\alias{rCalc}
\alias{rCalc,lhm-method}
\alias{rCalc,lhmIter-method}
\title{Calculate the intrinsic growth rate}
\usage{
rCalc(.Object, ...)

\S4method{rCalc}{lhm}(.Object)

\S4method{rCalc}{lhmIter}(.Object)
}
\arguments{
\item{.Object}{an \code{lhm} object}

\item{...}{additional arguments to generic function}
}
\value{
A \code{\link{prior}} class object containing a vector of values for \eqn{r}.
}
\description{
A value for the intrinsic growth rate \eqn{r} is calculated using life-history data as a numerical solution to the Euler-Lotka equation.
}
\examples{
# initialise lhm object
ainf <- 100
iter <- 1
rdat <- lhm(ainf,iter)

# assign life-history data
nmort(rdat)    <- list(mu = 0.18)
size(rdat)     <- list(mu = list(Linf = 106.5, k = 0.229, t0 = 0.01)) 
mass(rdat)     <- list(mu = list(a = 1.7e-9, b = 3.328))
sr(rdat)       <- list(type = 'BH', mu= 0.9)
maturity(rdat) <- list(mu = list(acrit = 8))

# calculate r
r <- rCalc(rdat)

}
